v1 rendering (#2)
Implemented Vulkan-based rendering with an interactive start/stop buttom and a realtime(ish) simulation display

* Implemented a first attempt at a vulkan window

* Implemented a basic main loop, and device selection

* Implemented logical device creation

* [SEGFAULTS] First pass at getting swapchain, graphics and present queues.

* [RUNS] Fixed segfault by not creating null QueueCreateInfos. Made debug info less verbose

* Implemented swapchain creation

* Implemented swapchain imageview getting, fixed build error in release mode

* Implemented basic shader compilation in CMake

* Improved structure of glsl CMake

* Implemented basic Vulkan shader loader

* [NOT BUILDING] Implemented templated shaders and pipelines/pipeline sets. Just needs a render pass to work.

* Added a renderpass and fixed up the pipelines

* Implemented command buffers for each frame

* [NOT WORKING] Theoretical rendering main_loop

* [WORKING] Got first triangle on-screen!

* Fixed destroy-while-in-use bug

* Got ImGUI to work

* Implemented custom RenderPass class to make creating new RenderPasses easier

* [BREAKS VALIDATION] First pass at multithreaded ImGUI

* Fixed validation errors in ImGUI

* Stopped using an exception to join threads

* Reorganized threading classes into separate files properly

* Implemented first shot at 2D allocator interface

* Added more verbose error checks to allocation

* Implemented allocator across all backends (only used in CUDA at the moment)
Templated Cuda backend and utils on "using Unified Memory or not" in prep for Vulkan memory
Changed per-dimension size type to uint32_t in non-vulkan scenarios

* Created consistent allocator pattern for memory that could be owner-visible

* Made allocate2D_unsafe protected so it could only be called from known-safe contexts.

* Started implementing Vulkan allocators for simulations

* Implemented CUDAVulkan2DAllocator, VulkanSimulationAllocator. Renamed CUDA2DUnifiedAllocator to be consistent with other allocators.

* [CUDA FAILS LINK] Added CUDA test to VulkanWindow in order to test the CUDA <-> Vulkan memory

* [CUDA FAILS BUILD] Implemented extract_data for non-unified CUDA arrays

* Made CUDA work with Unified arrays where possible, to make it finally compile. It's unusably slow tho

* Made CudaVulkan2DAllocator actually allocate device-local memory

* Made BaseVulkan2DAllocator defer initialization of buffers to the subclass after mapping, instead of ignoring initialization entirely.

* Added compile time checks to dispatch_gpu_prefetch to prevent it from being used on non-unified memory. This required making it nonvirtual, which is fine as it's never downcased.

* Added CUDA error check macros

* Made fatal_error use exceptions so that GDB breaks at the point where the error occurs

* Moved SystemWorker ImGui work outside of command buffer dispatch

* Implemented a simulation running in parallel with Vulkan

* Cleaned up printfs added in the last commit

* Added more shaders in preparation for sim display, added FPS count in ImGUI

* Implemented DescriptorSetLayouts and DescriptorSets for simulation

* Implemented simulation rendering.

* Separated the window dimensions from the simulation dimensions

* Made simulation display inside an ImGui window

+++ project/framework_test/src/simulation/file_format/SimSize.h
@@ -8,7 +8,7 @@
 #include "util/Size.h"
 
 struct SimSize {
-    SimSize(Size<size_t> pixel_size, Size<float> physical_size)
+    SimSize(Size<uint32_t> pixel_size, Size<float> physical_size)
         : pixel_size(pixel_size),
           physical_size(physical_size)
     {}
@@ -20,7 +20,7 @@ struct SimSize {
     }
 
     // Size of the simulation grid in pixels
-    const Size<size_t> pixel_size;
+    const Size<uint32_t> pixel_size;
     // Size of the simulation grid in meters
     const Size<float> physical_size;
